Added varying colours for the annotations
The colours are purely for demonstration, feel free to change them to
ones that look better.

Don’t know if this is a good change, just something I thought of and
whipped up real quick. Combined with a colour picker in the form
(there’s decent libraries for that) or a drop-down box with options, I
think it could work quite well. If you all disagree though, feel free
to delete it, I won’t be offended!